+++ raspberryrequest/modules/__init__.py
@@ -1,2 +1,3 @@
-from raspberryrequest.modules.calculate_backoff import calculate_backoff
-from raspberryrequest.modules.valid_status import valid_status
+from backoff import calculate_backoff
+from validate import valid_status
+from request import make_request